# packages/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LuckyScratch Backend API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/gasless/purchase:
    post:
      summary: Submit a gasless purchase request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequest"
      responses:
        "202":
          description: Accepted
  /api/v1/gasless/purchase-selection:
    post:
      summary: Submit a gasless purchase-selection request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseSelectionRequest"
      responses:
        "202":
          description: Accepted
  /api/v1/gasless/scratch:
    post:
      summary: Submit a gasless scratch request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScratchRequest"
      responses:
        "202":
          description: Accepted
  /api/v1/gasless/batch-scratch:
    post:
      summary: Submit a gasless batch-scratch request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchScratchRequest"
      responses:
        "202":
          description: Accepted
...
components:
  schemas:
    GaslessRequest:
      type: object
      required:
        - user
        - action
        - targetContract
        - paramsHash
        - nonce
        - deadline
        - chainId
      properties:
        user:
          type: string
        action:
          oneOf:
            - type: string
            - type: integer
        targetContract:
          type: string
        paramsHash:
          type: string
        nonce:
          oneOf:
            - type: string
            - type: integer
        deadline:
          oneOf:
            - type: string
            - type: integer
        chainId:
          oneOf:
            - type: string
            - type: integer
    PurchaseRequest:
      type: object
      required:
        - request
        - signature
        - poolId
        - quantity
      properties:
        request:
          $ref: "#/components/schemas/GaslessRequest"
        signature:
          type: string
        poolId:
          oneOf:
            - type: string
            - type: integer
        quantity:
          type: integer
    PurchaseSelectionRequest:
      type: object
      required:
        - request
        - signature
        - poolId
        - ticketIndexes
      properties:
        request:
          $ref: "#/components/schemas/GaslessRequest"
        signature:
          type: string
        poolId:
          oneOf:
            - type: string
            - type: integer
        ticketIndexes:
          type: array
          items:
            type: integer
    ScratchRequest:
      type: object
      required:
        - request
        - signature
        - ticketId
      properties:
        request:
          $ref: "#/components/schemas/GaslessRequest"
        signature:
          type: string
        ticketId:
          oneOf:
            - type: string
            - type: integer
    BatchScratchRequest:
      type: object
      required:
        - request
        - signature
        - ticketIds
      properties:
        request:
          $ref: "#/components/schemas/GaslessRequest"
        signature:
          type: string
        ticketIds:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
...